+++ src/CentroidCompletionProvider.ts
@@ -1,0 +1,20 @@
+"use strict";
+import * as vscode from "vscode";
+import { DocumentSymbolManager } from "./DocumentManager";
+import { getWordForPosition } from "./util";
+
+export class CentroidCompletionProvider
+  implements vscode.CompletionItemProvider {
+  provideCompletionItems(
+    document: vscode.TextDocument,
+    position: vscode.Position,
+    token: vscode.CancellationToken,
+    context: vscode.CompletionContext
+  ): vscode.ProviderResult<vscode.CompletionItem[] | vscode.CompletionList> {
+    let wordText = getWordForPosition(document, position);
+    let tries = DocumentSymbolManager.getTriesForDocument(document);
+    if (!tries) return [];
+    let symbolResults = tries.getAllCompletions(!wordText ? "" : wordText);
+    return symbolResults;
+  }
+}
